samc

+++ src/main/java/me/jweissen/aeticket/dto/response/CartEntryResponseDto.java
@@ -1,4 +1,4 @@
 package me.jweissen.aeticket.dto.response;
 
-public record CartEntryResponseDto(Integer id, String name, Double price, Integer amount) {
+public record CartEntryResponseDto(Long id, String name, Double price, Integer amount) {
 }
